Feature/ab2d 1273 (#279)
Add a button to start a bulk export that can optionally use a contract. Add a cancel button that can cancel an in progress export. On the backend add cross origin configuration so that headers can be sent and received via cross origin requests. Add a progress bar at the bottom that polls every 6 seconds to check on the status.