[AB2D] Update README.md with setup instructions (#61)
update the readme with setup instructions

## 🎫 Ticket

Unticketed

## 🛠 Changes

Updates the readme.md with setup instructions and instructions for
running the website locally.

## ℹ️ Context

These changes were done to make improve the readme, making it more clear
how to setup and run the project locally.

## 🧪 Validation

N/A - This is just a readme update, there are no functional updates to
the project. I edited the readme directly on github and committed to a
new branch.