[AB2D-6208] Add github action (#82)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6208

## 🛠 Changes

Adds a GitHub Action to manually deploy to development environment.
Specific updates for Production deployments will come in a separate PR

## ℹ️ Context

Automates the deployment of deploying to development to ensure more
consistent releases and to remove the need to manually share credentials
to specific users in order to deploy manually.

## 🧪 Validation

Manual QA