Implement privacy settings page (#94)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6486

## 🛠 Changes

NOTE: The scope of this work is front end only. The feature will not be
functional until we have working integration with metrics tools

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

Co-authored-by: Nat Watkins <watkinsnbw@gmail.com>