add privacy settings and sitemap to identifier (#117)
these are internal links and can't be added to the yml file

## 🎫 Ticket

Ad hoc spike 

## 🛠 Changes

Only change was the additional of two links to the identifier: `Sitemap`
and `Privacy Settings`. These can't be entered in the yml file because
they are internal links and/or have specific onclick values. The rest of
changes are formatting.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

---------

Co-authored-by: Jonathan Hutchison <jonathanhutchison@navapbc.com>