Fixes privacy modal trigger (#131)
## 🎫 Ticket


https://jira.cms.gov/browse/BLSTANALYT-11817?focusedId=4093515&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-4093515

## 🛠 Changes

Changes privacy modal trigger from a link to a button. Removes an unused
file.

## ℹ️ Context for reviewers

Modal trigger was a link element, which caused a "external_link" event
to be tracked in our analytics. Its also semantically incorrect.
Changing it to a button is both correct and prevents that event from
being attached to the trigger.

## ✅ Acceptance Validation

Manual QA

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.